book/requirements.txt
Django==3.2.13
django-environ==0.8.1
django-cors-headers==3.12.0
drf-yasg==1.20.0
django-filter==21.1
djangorestframework==3.13.1